Moves ast conversion to lang, and plan to its own lib
