+++ partiql-cli/build.gradle.kts
@@ -20,8 +20,8 @@ plugins {
 
 dependencies {
     implementation(project(":partiql-lang"))
+    implementation(project(":partiql-plan"))
     implementation(project(":partiql-types"))
-    implementation(project(":partiql-lib:partiql-plan"))
     implementation(Deps.csv)
     implementation(Deps.awsSdkBom)
     implementation(Deps.awsSdkDynamodb)
