+++ partiql-lang/build.gradle.kts
@@ -25,6 +25,7 @@ plugins {
 dependencies {
     antlr(Deps.antlr)
     api(project(":partiql-lib:partiql-isl"))
+    implementation(project(":partiql-plan"))
     implementation(project(":partiql-types"))
     api(Deps.ionElement)
     api(Deps.ionJava)
