+++ partiql-lang/src/main/kotlin/org/partiql/lang/planner/transforms/plan/RelConverter.kt
@@ -1,16 +1,16 @@
-package org.partiql.plan
+package org.partiql.lang.planner.transforms.plan
 
 import com.amazon.ionelement.api.ionInt
 import com.amazon.ionelement.api.ionString
 import org.partiql.lang.domains.PartiqlAst
 import org.partiql.lang.eval.visitors.VisitorTransformBase
-import org.partiql.plan.ir.Binding
-import org.partiql.plan.ir.Case
-import org.partiql.plan.ir.Common
-import org.partiql.plan.ir.Field
-import org.partiql.plan.ir.Rel
-import org.partiql.plan.ir.Rex
-import org.partiql.plan.ir.SortSpec
+import org.partiql.plan.Binding
+import org.partiql.plan.Case
+import org.partiql.plan.Common
+import org.partiql.plan.Field
+import org.partiql.plan.Rel
+import org.partiql.plan.Rex
+import org.partiql.plan.SortSpec
 import org.partiql.types.StaticType
 
 /**
