--- partiql-lib/partiql-plan/src/main/kotlin/org/partiql/plan/passes/notes.txt
@@ -1,8 +1,0 @@
-- Function Lowering
-- Constant Folding
-- Reduce Logical Expressions
-- Remove always true filters
-- Replace always false filters with empty query
-- Proj/Pred Pushdown
-
-Determine appropriate ordering
