+++ partiql-plan/build.gradle.kts
@@ -4,9 +4,7 @@ plugins {
 }
 
 dependencies {
-    // only for the org.partiql.domains package
-    implementation(project(":partiql-lang"))
     implementation(project(":partiql-types"))
-    // pretty printing
+    implementation(Deps.ionElement)
     implementation(Deps.kotlinReflect)
 }
