+++ partiql-plan/src/main/kotlin/org/partiql/plan/Types.kt
@@ -1,41 +1,41 @@
-package org.partiql.plan.ir
+package org.partiql.plan
 
 import com.amazon.ionelement.api.IonElement
-import org.partiql.plan.ir.builder.ArgTypeBuilder
-import org.partiql.plan.ir.builder.ArgValueBuilder
-import org.partiql.plan.ir.builder.BindingBuilder
-import org.partiql.plan.ir.builder.BranchBuilder
-import org.partiql.plan.ir.builder.CommonBuilder
-import org.partiql.plan.ir.builder.FieldBuilder
-import org.partiql.plan.ir.builder.PartiQlPlanBuilder
-import org.partiql.plan.ir.builder.RelAggregateBuilder
-import org.partiql.plan.ir.builder.RelBagBuilder
-import org.partiql.plan.ir.builder.RelFetchBuilder
-import org.partiql.plan.ir.builder.RelFilterBuilder
-import org.partiql.plan.ir.builder.RelJoinBuilder
-import org.partiql.plan.ir.builder.RelProjectBuilder
-import org.partiql.plan.ir.builder.RelScanBuilder
-import org.partiql.plan.ir.builder.RelSortBuilder
-import org.partiql.plan.ir.builder.RelUnpivotBuilder
-import org.partiql.plan.ir.builder.RexAggBuilder
-import org.partiql.plan.ir.builder.RexBinaryBuilder
-import org.partiql.plan.ir.builder.RexCallBuilder
-import org.partiql.plan.ir.builder.RexCollectionArrayBuilder
-import org.partiql.plan.ir.builder.RexCollectionBagBuilder
-import org.partiql.plan.ir.builder.RexIdBuilder
-import org.partiql.plan.ir.builder.RexLitBuilder
-import org.partiql.plan.ir.builder.RexPathBuilder
-import org.partiql.plan.ir.builder.RexQueryCollectionBuilder
-import org.partiql.plan.ir.builder.RexQueryScalarPivotBuilder
-import org.partiql.plan.ir.builder.RexQueryScalarSubqueryBuilder
-import org.partiql.plan.ir.builder.RexSwitchBuilder
-import org.partiql.plan.ir.builder.RexTupleBuilder
-import org.partiql.plan.ir.builder.RexUnaryBuilder
-import org.partiql.plan.ir.builder.SortSpecBuilder
-import org.partiql.plan.ir.builder.StepKeyBuilder
-import org.partiql.plan.ir.builder.StepUnpivotBuilder
-import org.partiql.plan.ir.builder.StepWildcardBuilder
-import org.partiql.plan.ir.visitor.PlanVisitor
+import org.partiql.plan.builder.ArgTypeBuilder
+import org.partiql.plan.builder.ArgValueBuilder
+import org.partiql.plan.builder.BindingBuilder
+import org.partiql.plan.builder.BranchBuilder
+import org.partiql.plan.builder.CommonBuilder
+import org.partiql.plan.builder.FieldBuilder
+import org.partiql.plan.builder.PartiQlPlanBuilder
+import org.partiql.plan.builder.RelAggregateBuilder
+import org.partiql.plan.builder.RelBagBuilder
+import org.partiql.plan.builder.RelFetchBuilder
+import org.partiql.plan.builder.RelFilterBuilder
+import org.partiql.plan.builder.RelJoinBuilder
+import org.partiql.plan.builder.RelProjectBuilder
+import org.partiql.plan.builder.RelScanBuilder
+import org.partiql.plan.builder.RelSortBuilder
+import org.partiql.plan.builder.RelUnpivotBuilder
+import org.partiql.plan.builder.RexAggBuilder
+import org.partiql.plan.builder.RexBinaryBuilder
+import org.partiql.plan.builder.RexCallBuilder
+import org.partiql.plan.builder.RexCollectionArrayBuilder
+import org.partiql.plan.builder.RexCollectionBagBuilder
+import org.partiql.plan.builder.RexIdBuilder
+import org.partiql.plan.builder.RexLitBuilder
+import org.partiql.plan.builder.RexPathBuilder
+import org.partiql.plan.builder.RexQueryCollectionBuilder
+import org.partiql.plan.builder.RexQueryScalarPivotBuilder
+import org.partiql.plan.builder.RexQueryScalarSubqueryBuilder
+import org.partiql.plan.builder.RexSwitchBuilder
+import org.partiql.plan.builder.RexTupleBuilder
+import org.partiql.plan.builder.RexUnaryBuilder
+import org.partiql.plan.builder.SortSpecBuilder
+import org.partiql.plan.builder.StepKeyBuilder
+import org.partiql.plan.builder.StepUnpivotBuilder
+import org.partiql.plan.builder.StepWildcardBuilder
+import org.partiql.plan.visitor.PlanVisitor
 import org.partiql.types.StaticType
 import kotlin.Any
 import kotlin.String
